Revert "Test: visitAuth 디자인 불러오기"
