+++ catch-culture-frontend/package.json
@@ -12,7 +12,6 @@
   "dependencies": {
     "@react-icons/all-files": "^4.1.0",
     "axios": "^1.6.2",
-    "http-proxy-middleware": "^2.0.6",
     "pagination": "^0.4.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -30,6 +29,5 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.3.4",
     "vite": "^4.3.9"
-  },
-  "proxy": "https://elegant.kro.kr"
+  }
 }
