+++ catch-culture-frontend/src/api/axios.js
@@ -1,8 +1,7 @@
 import axios from 'axios';
 
 const instance = axios.create({
-  baseURL: 'https://elegant.kro.kr/',
-  withCredentials: true,
+  baseURL: 'http://elegant.kro.kr/',
   headers: {
     'Content-Type': 'application/json',
   },
