+++ catch-culture-frontend/src/components/categorySelector/categorySelectorItem/CategorySelectorItem.jsx
@@ -8,19 +8,19 @@ function CategorySelectorItem(props) {
         onClick={() => props.handlerClickCategory(props.category)}
         color={
           props.selectedCategories.includes(props.category) ||
-          (props.category === 'ALL' && props.selectedCategories.length === 13)
+          (props.category === 'ALL' && props.selectedCategories.length === 12)
             ? '#fff'
             : '#018C0D'
         }
         bgcolor={
           props.selectedCategories.includes(props.category) ||
-          (props.category === 'ALL' && props.selectedCategories.length === 13)
+          (props.category === 'ALL' && props.selectedCategories.length === 12)
             ? '#018C0D'
             : '#fff'
         }
         shadow={
           props.selectedCategories.includes(props.category) ||
-          (props.category === 'ALL' && props.selectedCategories.length === 13)
+          (props.category === 'ALL' && props.selectedCategories.length === 12)
             ? '0px 4px 4px 0px rgba(0, 0, 0, 0.25) inset'
             : 'none'
         }
