+++ catch-culture-frontend/src/main.jsx
@@ -5,5 +5,7 @@ import router from './router.jsx';
 import './assets/fonts/Font.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
+  <React.StrictMode>
     <RouterProvider router={router} />
+  </React.StrictMode>
 );
