--- catch-culture-frontend/src/pages/culturalEventDetail/eventReview/style.jsx
@@ -1,32 +1,0 @@
-import { styled } from 'styled-components';
-
-export const EventInfo = styled.div`
-    text-align: center;
-
-    display: flex;
-    align-items: center;
-    justify-items: center;
-    flex-direction: column;
-`;
-
-export const TitleArea = styled.div`
-    width: 200px !important;
-    height: 38px;
-    text-align: center;
-    font: 20px bold;
-    font-weight: bold;
-    font-family: NotoSansKR_Bold;
-    margin-top: 20px;
-    margin-bottom: 10px;
-    border-bottom: 2px solid black;
-`;
-
-export const CategoryArea = styled.div`
-    width: 40px;
-    height: 12px;
-    font-size: 8px;
-`;
-
-export const AuthArea = styled.div`
-    font-size: 12px;
-`;
