+++ catch-culture-frontend/src/pages/main/Main.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 import * as S from './style';
 import LogoImg from '../../assets/images/logo.png';
@@ -15,29 +15,7 @@ import 'swiper/css';
 import 'swiper/css/pagination';
 import { Pagination } from 'swiper/modules';
 
-// api
-import axios from '../../api/axios';
-
 function Main() {
-  // data
-  const [data, setData] = useState([]);
-
-  // 초기
-  useEffect(() => {
-    fetchData();
-  }, []);
-
-  const fetchData = async () => {
-    try {
-      const response = await axios.get('cultural-event');
-
-      //데이터 저장
-      setData(response.data);
-    } catch (e) {
-      console.log(e);
-    }
-  };
-
   return (
     <>
       <S.MainWrapper>
@@ -74,7 +52,7 @@ function Main() {
               <S.EventHeaderMore>더보기</S.EventHeaderMore>
             </Link>
           </S.EventHeader>
-          <EventCard data={data} />
+          <EventCard />
         </S.EventWrapper>
       </S.MainWrapper>
     </>
