+++ catch-culture-frontend/src/pages/mypage/BookmarkList.css
@@ -1,42 +1,8 @@
-.tota {
+.bookmarktext {
   display: flex;
-  position: relative;
-  gap: 40px;
-}
-
-.wrap {
-  display: flex;
-  position: relative;
-  flex-direction: column;
-  gap: 20px;
-}
-
-.cateSel {
-  position: relative;
-  display: flex;
-}
-
-.textrow {
-  display: flex;
-  position: relative;
-  align-items: baseline;
-  gap: 8px;
-}
-.bookmarkCnt {
-  font-size: 12px;
-}
-.liketext {
+  position: absolute;
+  top: 8%;
+  margin: 0 0 0 -176px;
   font-size: 20px;
   font-weight: bold;
-}
-
-.eventlist {
-  display: flex;
-  position: relative;
-}
-.nors {
-  display: flex;
-  position: relative;
-  width: 100%;
-  justify-content: center;
 }
