--- catch-culture-frontend/src/pages/mypage/Reviews.css
@@ -1,17 +1,0 @@
-.reviewrow {
-  display: flex;
-  gap: 8px;
-  width: 322px;
-  align-items: baseline;
-}
-.reviewtext {
-  font-size: 20px;
-  font-weight: bold;
-}
-.reviewCnt {
-  font-size: 12px;
-}
-.reviewcard {
-  display: flex;
-  justify-content: center;
-}
