--- catch-culture-frontend/src/pages/reportPage/Report.jsx
@@ -1,8 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import * as S from './style';
-
-function Report() {
-  return <></>;
-}
-
-export default Report;
