+++ catch-culture-frontend/src/pages/search/Search.jsx
@@ -15,7 +15,6 @@ import axios from '../../api/axios';
 function Search() {
   const { state } = useLocation();
   const category = state && state.category;
-  const [count, setCount] = useState(0);
 
   console.log(category);
 
@@ -39,6 +38,9 @@ function Search() {
   // data
   const [data, setData] = useState([]);
 
+  // 임시 cnt
+  const cnt = 2;
+
   // 카테고리 바뀔 때 마다 리렌더링
   useEffect(() => {
     // console.log(selectedCategories);
@@ -53,22 +55,17 @@ function Search() {
 
   const fetchData = async () => {
     try {
-      if (selectedCategories.length === 0) {
-        setCount(0);
-      } else {
-        // URL 만들기 - 카테고리 선택
-        const categoryUrl = selectedCategories
-          .map(item => 'category=' + item)
-          .join('&');
+      // URL 만들기 - 카테고리 선택
+      const categoryUrl = selectedCategories
+        .map(item => 'category=' + item)
+        .join('&');
 
-        const response = await axios.get(
-          `cultural-event/list?${categoryUrl}&offset=0&sortType=${options[selectedSort].label}`
-        );
+      const response = await axios.get(
+        `cultural-event?${categoryUrl}&offset=0&sortType=${options[selectedSort].label}`
+      );
 
-        // 데이터 저장
-        setData(response.data.content);
-        setCount(response.data.totalElements);
-      }
+      // 데이터 저장
+      setData(response.data);
     } catch (e) {
       console.log(e);
     }
@@ -83,7 +80,7 @@ function Search() {
         {/* 헤더 */}
         <S.SearchHeader>
           <S.SearchHeaderTitle>검색 결과</S.SearchHeaderTitle>
-          <S.SearchHeaderResultCnt>총 {count}개</S.SearchHeaderResultCnt>
+          <S.SearchHeaderResultCnt>총 {cnt}개</S.SearchHeaderResultCnt>
         </S.SearchHeader>
 
         {/* 카테고리 선택창 */}
@@ -102,7 +99,7 @@ function Search() {
         </S.SortSelectorWrapper>
 
         {/* 문화 행사 출력 */}
-        {count === 0 ? (
+        {cnt === 0 ? (
           <>
             <NoResult />
           </>
